+++ resources/js/Components/Shows/ShowFooter.vue
@@ -1,6 +1,6 @@
 <template>
 
-    <div class="flex justify-end mt-6 mb-24">
+    <div class="flex justify-end mt-6 pb-24">
         <!-- Paginator -->
         <!--                            <Pagination :links="`#`" class="mt-6"/>-->
         <Link :href="`/teams/${team.slug}`" class="text-blue-500 ml-2 uppercase"> {{ team.name }} ©